TST: Reorganizing test_tools_vector.py
